Add SalesReportsLoad
Intoduced FilesSystemUtils project to store classes for reading and writing certian file types to the file system.  Implemented Composite pattern via IArchiveLoader that has many IFileLoaders, so simultaneous loading of various file types with single archive read can be made. You can even read archives inside archives and so on.

In Logic assembly added DataLoaders namespace to store various loaders of data to a destination DbContext. Added SalesReportsLoader to write to FurnitureDataDbContext.Clients. So far only the new clients that are read from an archive are added to Clients DbSet and saved to the database.