Finished SalesReports importing
SalesReportImporter.ImportData()
The method imports data to the sql server database. It checks whether such customer exists
 and adds it to customers DbSet if necessary otherwise gets its id to use for orders.
 So you can`t add the same customer twice to Customers DbSet, say if you load two different xls reports from the same customer.
But it does not check for duplication of orders.
So the same orders can be added to Orders DbSet twice if we load the same xls report file twice from a particular Customer.